+++ src/datasets/Datasets.jl
@@ -9,27 +9,16 @@ import MLDatasets
 using MLUtils: mapobs, groupobs
 using DataDeps
 using FilePathsBase
-import DataAugmentation
 using FilePathsBase: filename
 import FileIO
-using IndirectArrays: IndirectArray
-using Colors
-using FixedPointNumbers
-using DataFrames
-using Tables
-using CSV
-using ShowCases
 using InlineTest
 
 include("fastaidatasets.jl")
 
 include("batching.jl")
-include("containers.jl")
 include("load.jl")
 include("loaders.jl")
 include("recipe.jl")
-
-include("deprecations.jl")
 
 function __init__()
     initdatadeps()
@@ -37,16 +26,11 @@ end
 
 export
     # primitive containers
-    TableDataset,
-
     mapobs, eachobs, groupobs, shuffleobs, ObsView,
 
     # utilities
-    isimagefile,
-    istextfile,
     matches,
     loadfile,
-    loadmask,
     pathname,
     pathparent,
     parentname,
@@ -54,12 +38,7 @@ export
 
     # datasets
     loadfolderdata,
-    datasetpath,
+    datasetpath
 
-    # recipes
-    loadrecipe,
-    finddatasets,
-    listdatasources,
-    loaddataset
 
 end  # module